most of the time attachment is Long
git-svn-id: https://svn.java.net/svn/grizzly~svn/branches/grizzly1.0@3851 04fdb32c-e91d-0410-84f5-98e489d73e21

+++ src/main/java/com/sun/enterprise/web/portunif/PortUnificationPipeline.java
@@ -268,7 +268,7 @@ public class PortUnificationPipeline extends SSLPipeline{
                  * Since we can't predict what the user attach to the SelectionKey
                  * a ClassCastException is required.
                  */
-                Integer maxKeepAliveRequestsToken = (Integer)key.attachment();
+                Long maxKeepAliveRequestsToken = (Long) key.attachment();
                 if (maxKeepAliveRequestsToken == -1
                         && ph.getClass().getName().equals("com.sun.xml.ws.transport.tcp.grizzly.WSTCPProtocolHandler")){
                     return false;
